JENKINS-53675 Moved and renamed the suppression annotation
SuppressRestrictedWarnings mirrors java.lang.SuppressWarnings naming
Moved to a separate package makes it less-available to users and puts more
barriers in for use. This makes the user stop-and-think about what they
are doing, rather than just turning the warning off.
