+++ access-modifier-checker/src/it/disable-restrictions-wrong-annotation/caller/src/main/java/caller/CallerDisabledAtClassLevel.java
@@ -1,8 +1,6 @@
 package caller;
 
 import api.ApiWithRestrictedMethodAndField;
-import api.RestrictedApi;
-import org.kohsuke.accmod.restrictions.disable.DisableRestriction;
 
 @WrongAnnotation(ApiWithRestrictedMethodAndField.class)
 public class CallerDisabledAtClassLevel {
