+++ access-modifier-checker/src/it/disable-restrictions-wrong-annotation/postbuild.groovy
@@ -1,2 +1,2 @@
 assert new File(basedir, 'build.log').text.contains('[ERROR] caller/Caller:9 api/ApiWithRestrictedMethodAndField.notReallyPublic()V must not be used')
-assert new File(basedir, 'build.log').text.contains('[ERROR] caller/CallerDisabledAtClassLevel:10 api/ApiWithRestrictedMethodAndField.notReallyPublic()V must not be used')
+assert new File(basedir, 'build.log').text.contains('[ERROR] caller/CallerDisabledAtClassLevel:8 api/ApiWithRestrictedMethodAndField.notReallyPublic()V must not be used')
