+++ access-modifier-checker/src/it/disable-restrictions/caller/src/main/java/caller/Caller.java
@@ -2,34 +2,34 @@ package caller;
 
 import api.ApiWithRestrictedMethodAndField;
 import api.RestrictedApi;
-import org.kohsuke.accmod.restrictions.disable.DisableRestriction;
+import org.kohsuke.accmod.restrictions.suppressions.SuppressRestrictedWarnings;
 
 public class Caller extends ApiWithRestrictedMethodAndField { // This is fine, ApiWithRestrictedMethodAndField itself is not restricted
 
     private RestrictedApi restrictedApi;
 
-    @DisableRestriction(ApiWithRestrictedMethodAndField.class)
+    @SuppressRestrictedWarnings(ApiWithRestrictedMethodAndField.class)
     public Caller() {
         ApiWithRestrictedMethodAndField.notReallyPublic(); // illegal but check disabled at the method level
     }
 
-    @DisableRestriction({RestrictedApi.class, ApiWithRestrictedMethodAndField.class})
+    @SuppressRestrictedWarnings({RestrictedApi.class, ApiWithRestrictedMethodAndField.class})
     private void invalidFieldUse() {
         restrictedApi.field = null;
         super.field = null;
     }
 
-    @DisableRestriction(ApiWithRestrictedMethodAndField.class)
+    @SuppressRestrictedWarnings(ApiWithRestrictedMethodAndField.class)
     public void callerMethod() {
         ApiWithRestrictedMethodAndField.notReallyPublic(); // illegal but check disabled at the method level
     }
 
-    @DisableRestriction(RestrictedApi.class)
+    @SuppressRestrictedWarnings(RestrictedApi.class)
     public void methodWithRestrictedParameter(RestrictedApi api) {
         api.doNotUse(); // illegal but check disabled at the method level
     }
 
-    @DisableRestriction(RestrictedApi.class)
+    @SuppressRestrictedWarnings(RestrictedApi.class)
     public RestrictedApi getRestrictedApi() {
         return new RestrictedApi(); // illegal but check disabled at the method level
     }
